Fix: properly save timestamps with timezones
Previous test cases only covered +00:00 as timezone. Adding test cases
for +01:00 resulted in data getting returned with the "same" timestamp,
but in the wrong timezone (+00:00). This means you lost an hour in the
conversion.

Because the Excel timestamp format does not seem to store timezones, we
might as well convert everything to +00:00 / UTC and do a proper
conversion.

Because Date/DateTime and Time have different APIs, they use their own
method to convert to +00:00.